% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dummy variable.R
\name{factor}
\alias{factor}
\title{Generate Dummy Variables for Macrox Analysis}
\usage{
factor(typename, d_v = NULL)
}
\arguments{
\item{typename}{Character. The name of the column for which dummy variables are to be created. Should be a single string representing the column name.}

\item{d_v}{Optional. A vector of values specifying which categories should be used to generate dummy variables.
If \code{d_v} is not provided, the function will automatically generate dummy variables based on the elements present in the \code{typename} column.}
}
\value{
A list containing:
\itemize{
\item A vector with the original \code{typename} and the names of the generated dummy variables.
\item A data frame of the generated dummy variables added to the original data.
}
}
\description{
This function generates dummy variables for specified categorical columns, typically used as part of the \code{formula} parameter within the \code{marcox()} function.
If \code{d_v} is not provided, dummy variables will be automatically created based on the unique values within the specified \code{typename} column.
}
\details{
This function is typically used to prepare categorical data for analysis within \code{marcox()}.
If the \code{d_v} parameter is omitted, the function will automatically determine the categories from the \code{typename} column and create dummy variables.
The generated dummy variables will be appended to the original data frame (\code{cluster2}).
If the \code{d_v} parameter is provided, it should specify which categories to convert to dummy variables.
If any \code{type} variables (categorical variables) are included, they must be passed as strings when using \code{factor()}.
}
