% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize Data for Macrox Analysis}
\usage{
init(ad, sep = "\\t", div = NULL)
}
\arguments{
\item{ad}{The file path or data frame to analyze. If a file path is provided, the file will be loaded into a data frame.
The file should be in a tabular format (e.g., .csv, .txt).}

\item{sep}{parameter. The \code{sep} parameter specifies the character that separates
the fields in each line of the file. For instance, for a comma-separated file, set \code{sep = ","},
and for a tab-separated file, set \code{sep = "\t"}.
If the data has complex observational situations, please preprocess the data before using this function.}

\item{div}{Integer. The number of observation points per sample. If provided, the data will be divided accordingly. Default is \code{NULL}.}
}
\value{
A list containing:
\itemize{
\item \code{cluster2} - The processed data frame after applying any transformations, including the addition of unique IDs.
\item \code{dv} - A list of dummy variables. If no dummy variables are created, returns an empty list.
\item \code{col_name_origin} - The original column names of the data frame.
}
}
\description{
This function initializes data for further analysis, typically used as a parameter within the \code{marcox} function.
It loads data from a specified file or data frame, performs basic preprocessing, and assigns unique identifiers for grouping.
The function allows for simple clustering based on the \code{div} parameter, but if the data has more complex observational structures,
users should preprocess the data before running this function.
}
\details{
This function is generally used as a preparatory step for the \code{marcox} function.
It handles loading and preprocessing data by setting up unique identifiers (\code{id}) based on the number of observation points specified by \code{div}.
If \code{div} is provided, the function will ensure consistent grouping across the data.
If the number of rows in the data is not divisible by \code{div}, an error will be shown.
For more complex observational settings, it is recommended to preprocess the data before running this function.
}
\examples{
# Use an existing data frame without specifying div
sample_data_1 <- kidney_data
init(sample_data_1,div=2)
}
