% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat.R
\name{gendat}
\alias{gendat}
\title{Generate Datasets for Cox Proportional Hazards Model Simulation}
\usage{
gendat(
  dimension = 10,
  K = 30,
  n = 2,
  lambda = c(1, 2),
  b1 = log(2),
  theta = 5,
  censrate = 0.5,
  result_data_length = NULL
)
}
\arguments{
\item{dimension}{Integer. The number of datasets to be generated.}

\item{K}{Integer. The number of samples within each cluster.}

\item{n}{Integer. The number of clusters (groups) within each dataset.}

\item{lambda}{Numeric vector. A vector of two values specifying the parameters for the baseline distribution.
If \code{lambda = c(2,2)}, the baseline follows a Weibull distribution; if \code{lambda = c(1,2)}, the baseline follows an exponential distribution.}

\item{b1}{Numeric. The regression coefficient for the covariate, affecting the hazard function.}

\item{theta}{Numeric. A parameter that controls the dependence structure between survival times within clusters.}

\item{censrate}{Numeric. The desired censoring rate for the dataset.}

\item{result_data_length}{Integer or NULL. If specified, truncates each dataset to the given number of rows.}
}
\value{
A list containing:
\itemize{
\item \code{data} - A list of data frames, each containing a generated dataset.
\item \code{censoringrates} - A numeric vector of censoring rates for each dataset.
\item \code{mean(censoringrates)} - The mean censoring rate across all datasets.
}
}
\description{
This function generates multiple datasets for survival analysis under a Cox proportional hazards model.
The baseline hazard function follows either a Weibull or an exponential distribution, depending on the values of \code{lambda}.
The function ensures that the maximum observed time in both the control and treatment groups is checked for censoring.
If the maximum time is not censored, it is forced to be censored to maintain the desired censoring rate.
Users can also specify \code{result_data_length} to control the length of each dataset.
After running the function, type \code{datasets} in the R console to access all the generated datasets and relevant information.
}
\examples{
# Generate datasets with 3 dimensions, 2 clusters, 100 samples per cluster
gendat(dimension = 5, K = 100, n = 2, lambda = c(1, 2), b1 = log(2), theta = 8, censrate = 0.3,result_data_length=20)
# Access the generated datasets
datasets
}
