% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcox.R
\name{marcox}
\alias{marcox}
\title{Analysis for Cox Proportional Hazards Models}
\usage{
marcox(formula, dat, method = "exc")
}
\arguments{
\item{formula}{A model formula that uses the \code{Surv()} function to define the survival outcome. It should include
both continuous and categorical covariates, where categorical variables must be specified using the \code{factormar()} function.}

\item{dat}{A list containing the dataset as prepared by the \code{init()} function. This list must include the processed data frame
and original column names to ensure proper matching of variables.}

\item{method}{The method employed to solve the correlation coefficient:
\itemize{
\item Exchangeable correlation structure: \code{method = 'exc'}
\item Autoregressive(AR-1): \code{method = 'ar1'}
\item k-dependent: \code{method = 'kd'}
\item Toeplitz: \code{method = 'toep'}
\item Independent: \code{method = 'indp'}
\item Unstructured: \code{method = 'uns'}
}}
}
\value{
A data frame containing the following components:
\itemize{

\if{html}{\out{<div class="sourceCode">}}\preformatted{  \\item \code{coef} - The estimated regression coefficients.
  \\item \code{exp(coef)} - The exponentiated coefficients (hazard ratios).
  \\item \code{se(coef)} - The standard errors of the estimated coefficients.
  \\item \code{z} - The z-statistics for testing the significance of the coefficients.
  \\item \code{p} - The p-values associated with the coefficients.
  \\item \code{Correlation]} - The correlation coefficient of the data.
}\if{html}{\out{</div>}}

}
}
\description{
This function performs marcox analysis for Cox proportional hazards models, incorporating clustered data
and handling time-dependent covariates. It estimates coefficients, standard errors, and p-values based on
the specified formula and dataset.
}
\details{
The \code{marcox()} function is specifically designed for survival data analysis using Cox proportional hazards models. It handles both clustered and time-dependent covariates effectively.
The survival outcome must be defined using the \code{Surv()} function in the model formula, and covariates can be included directly or by converting categorical variables with the \code{factormar()} function.
}
\examples{
  dat <- init(kidney_data, div = 2)
  formula <- Surv(time, cens) ~ sex + factormar('type', d_v=c(1,2,3))
  result1 <- marcox(formula, dat, method = 'exc')
  print(result1)
}
