% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcox.R
\name{marcox}
\alias{marcox}
\title{Macrox Analysis for Cox Proportional Hazards Models}
\usage{
marcox(formula, dat)
}
\arguments{
\item{formula}{A formula object specifying the model to be fitted. Must include a \code{Surv()} function to define the survival object.
Dummy variables can be created using the \code{factor()} function.
If any \code{type} variables (categorical variables) are included, they must be passed as strings when using \code{factor()}.}

\item{dat}{A list containing the dataset initialized by the \code{init()} function. This list should include the processed data frame and any original column names.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{coef} - The estimated coefficients for the covariates.
\item \code{exp(coef)} - The exponentiated coefficients, representing the hazard ratios.
\item \code{se(coef)} - The standard errors of the estimated coefficients.
\item \code{z} - The z-values for the coefficients.
\item \code{p} - The p-values associated with the coefficients.
}
}
\description{
This function performs Macrox analysis for Cox proportional hazards models, incorporating clustered data and handling time-dependent covariates.
It estimates coefficients, standard errors, and p-values based on the specified formula and dataset.
The \code{formula} parameter can include dummy variables created using the \code{factor()} function, and the \code{dat} parameter must be initialized using the \code{init()} function.
}
\details{
The \code{marcox()} function is used to analyze survival data, specifically for clustered or time-dependent scenarios.
The \code{formula} must be constructed using the \code{Surv()} function to properly define the time-to-event data, and can include covariates specified directly or created via the \code{factor()} function.
The \code{dat} parameter should be prepared using the \code{init()} function, which sets up the dataset with necessary variables and identifiers.

This function outputs the analysis results, including estimates, standard errors, z-values, and p-values, and provides a summary of the model fit.
}
\examples{
# Assuming `init` has been used to prepare the dataset
# dat <- init(kidney_data, div = 2)
# formula <- Surv(time,cens)~sex+factor('type')
# marcox(formula, dat)
}
