% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcox.R
\name{marcox}
\alias{marcox}
\title{Analysis for Cox Proportional Hazards Models}
\usage{
marcox(formula, dat)
}
\arguments{
\item{formula}{A model formula that uses the \code{Surv()} function to define the survival outcome. It should include
both continuous and categorical covariates, where categorical variables must be specified using the \code{factormar()} function.}

\item{dat}{A list containing the dataset as prepared by the \code{init()} function. This list must include the processed data frame
and original column names to ensure proper matching of variables.}
}
\value{
A data frame (if \code{penalize=FALSE}) or a list (if \code{penalize=TRUE}) containing the following components:
\itemize{
\item \strong{For the standard marcox analysis}:
\itemize{
\item \code{coef} - The estimated regression coefficients.
\item \code{exp(coef)} - The exponentiated coefficients (hazard ratios).
\item \code{se(coef)} - The standard errors of the estimated coefficients.
\item \code{z} - The z-statistics for testing the significance of the coefficients.
\item \code{p} - The p-values associated with the coefficients.
\item \code{Correlation]} - The correlation coefficient of the data.
}

}
}
\description{
This function performs marcox analysis for Cox proportional hazards models, incorporating clustered data
and handling time-dependent covariates. It estimates coefficients, standard errors, and p-values based on
the specified formula and dataset. In addition, an optional penalization functionality is provided via the
\code{penalize} parameter. When \code{penalize} is set to TRUE, the function applies a SCAD-type penalty with
tuning parameter selection via a GCV criterion to perform variable selection.
}
\details{
The \code{marcox()} function is specifically designed for survival data analysis using Cox proportional hazards models. It handles both clustered and time-dependent covariates effectively.
The survival outcome must be defined using the \code{Surv()} function in the model formula, and covariates can be included directly or by converting categorical variables with the \code{factormar()} function.

When the \code{penalize} parameter is enabled, the function applies a SCAD-type penalty to perform variable selection. A range of tuning parameters,
determined by \code{pen_tp}, is evaluated via a GCV criterion to select the optimal penalty strength. This process helps in reducing model complexity by
shrinking insignificant coefficients toward zero.
}
\examples{
  dat <- init(kidney_data, div = 2)
  formula <- Surv(time, cens) ~ sex + factormar('type', d_v=c(1,2,3))
  result1 <- marcox(formula, dat)
}
